Sheol: Fix Sheol-only monsters not generated in Sheol.
This was caused by a generation flag G_SHEOLONLY meaning that
a monster should not be generated outside Sheol. The trick is that
it also doesn't mean it should be generated in Sheol (that would need
G_SHEOL flag to be set on the monster as well). This commit makes
G_SHEOLONLY also include G_SHEOL.