Numerous minor changes, mostly adjustments to the Moria branch:
Lessen the number of orcs in the Halls of Moria.
Give the orc-captains in the Halls a 25% chance of carrying a lit oil lamp.
Make dwarves and deep orcs dislike each other.
Add an additional level to Moria to distance the West-gate from the rest of the Ruins.
Incorporate the tree cutting patch by Michael Lehotay (http://bilious.alt.org/?67).