Sheol: Remove passive blue slime attack.
Blue slimes have an active and a passive attack but the passive attack
apparently does not work. However, instead of making it work, I think it is
better that they don't have the attack, like it is now. This commit prevents
the passive attack from showing up in any (up-to-date) external collections
monster information (such as the Pinobot and the Nethackwiki).