Fix a bug in irregular room creation.
In the new level compiler, the room "furniture" - such as stairs, thrones
and altars - are created in the order they appear in the des-file.  This
means the room may get stocked with those before the room floor space is
determined by a flood filling algorithm.  The algorithm was checking
strict floor type, instead of checking IS_ROOM(), and this caused parts
of the room not actually be part of the room.