If I did this correctly, the horn of plenty has a small chance of producing blood or vampire blood.
The way I did it is a bit strange, but right now when it tries to replace a magical potion with a "mundane" potion, it has a chance to produce a potion of (vampire) blood instead.

The probability looks rather small, but that is because in theory the probability might be checked multiple times. It is probably too low though.