Show message when formally identifying whistles on use
e.g. When applying a tin whistle produces "a high whistling sound" the
identifying message given is "It must be a tin whistle!"

The new makeknown_msg() function should be used in a lot of other places
(with care given to message ordering), but for now whistles serve as a
good proof-of-concept for it.