Fix writing with fingertips from menu (reported by oh6)
Writing with your fingertip did not work when tried from the menu due to
'-' being the same character as used for unselecting everything. NetHack
fixed this in commit 45c5a79681. The proper way of fixing should
probably be to implement a special mode for menus when only picking a
single item.

Tty menu hack for ':'

Looting a container generates a menu which contains ': - look inside'
but the recent change to make ':' be a menu command for selecting
items which match a search string made it impossible to pick that item.
(Well, I suppose you could enter a search string which matched it, but
that's a nuisance compared to just directly picking a choice.)  This
makes menu selection for tty give precedence to menu choice characters
over mapped menu commands when some character happens to be both.  I'm
not sure whether it ought to be expended to group accelerators too, so
didn't do that.

There's bound to be a better way to do this, but it works.

Cherry picked from nethack/nethack@45c5a79681