Disable the rogue tribute level
The dungeon is getting full and the rogue tribute level is adding little
in gameplay variation and is mostly a nuisance for players or worse a
confusing level for new players that haven't played or even heard of
rogue.

After 35 years, we should be able to let go of the older, more primitive
world.