Monster generation rate increases slower and more linearly with turn count
The original monster generation algorithm was taken from SporkHack and
somewhat adjusted. It increased the monster generation rate with an
exponential rate, leading to a steep increase to the max rate within
10000 turns around turn 80000.

This commit changes it that the max rate is 7x the base rate.  Below the
castle, it is at least 2x, after killing the Wizard of Yendor it is the
max rate of 7x.

The base rate is the same as Vanilla's, a new monster is generated on
average every 70 turns.

In Vanilla, the rate is every 50 turns (1.4x) below the castle or every
25 turns (2.8x) after killing the Wizard.