Rename the Dark One to Anaraxis the Black
Mainly this is because "Dark One" is a really stupid name. It sounds
like something an edgy teenager would name themselves. He's supposed to
be sort of an expy of Sauron, or something, but even if I did want to
stick with that flavor (I don't) there would be many better things to
name him.

Angling towards the flavor of him being a powerful rival wizard, he
needs a color epithet, thus "the Black". The name "Anaraxis" isn't a
reference to anything.

As with Tiamat, this only changes his name and not any of his stats.