Optimized placement of menu windows in TTY port
Default placement of menu windows was right aligning them with the
terminals width.

This resulted in menu windows being way over on the right side if you
have a very wide terminal or if the terminal is wider than 255
characters, menu windows being left aligned (like the win_edge option)
due of a number overflow bug in calculating the x offset (C343-439).

This commit changes the calculation of the x offset such that it will
put the menu window right next to the map if there is enough horizontal
space. Otherwise it overlaps the window with the map as before.