Don't cache mtmp coordinates too soon in dogmove()
The mx/my coordinates of a monster could be modified (like with the
uranium imp's attack) and the remove_monster() call with the stale
omx/omy variables would then not clear the correct level.monsters index.
Together with a deallocated monster, there would be a pointer to cleared
and deallocated memory left in level.monsters.