Improvements to how healers can sense injuries.
A while ago I had implemented L's patch (modified, in EvilHack) that
gave healers a sense of how wounded their target was. I also added in a
bit where a healer could also sense how wounded their pets were. This
patch was recently introduced to xNetHack, and its implementation is
better in a couple different ways. It's a bit less spammy, meaning you
won't see damage feedback on every single hit, but only when the damage
level changes. It also fixes a bug with non-damaging wand attacks
displaying damage feedback. So a revamp was in order, and I've updated
the code here to closely match it.