Adjusted quantity formula when wishing for stackable items
With the limited number of wishes available in UnNetHack, there's a
rationale to make each wish a little more useful, especially for
non-magical wishes.

This commit changes the maximal obtainable quantity of stackable items
with a single wish according to the following rules:

When wishing for a quantity of stackable items, the program randomly
picks a maximum quantity that the player can wish for.

For general items, this is up to 10 items (with a mean value of 5), for
projectiles it is up to 100 with a minimum of about 20 projectiles.

It is always possible to wish for up to 7 candles.

If the player wishes for more items than the quantity the program has
picked, the player gets adjusted by their luck either the previously
picked quantity or just a single item.