Improve branch location placement
Lava caverns could have rooms that would be completely overflown by
lava. If find_branch_room() picked such a room, it wouldn't be able to
find a suitable location and emit an impossible "can't place branch"
warning.

The code picks now a room and verifies if there is a suitable location
in the same step.