Output "You begin awkwardly bashing" for unskilled but skillable weapons
The game should now always inform the player if they're using an
unsuitable weapon with "You begin bashing ...". This includes weapons
for which the player has no skills.

For objects that are suitable weapons (that is the player has a weapon
skill for it that can be advanced past unskilled in the future) the
message "You begin awkwardly bashing" will be output.

Inspired by copperwater/xnethack@302ff52f2 which got it from FIQHack.