Potion mixing needs equal numbers of potions on both sides
One inconsistent aspect of color alchemy was the vanilla dipping
requirements. If potion alchemy is based on color why would a single
potion be able to alchemize an arbitrary number of potions.

This commit changes the code to require the same number of potions for
both sides of the alchemy process. The same rationale is applied for the
number of alchemized potions be to combined total of the involved
potions.