Shopkeepers honor Pride Month
During Pride Month (i.e. June) shopkeepers with a suitable sized shop
will sort their inventory by color for a rainbow pattern.

Currently, mimics don't color align with the rainbow pattern and
obviously need more convincing in the future. It's probably a difficult
concept to grasp for something that can be whatever it wants to be.