Wielding Vorpal Blade protects against decapitation attacks.
For a long time I thought this was already a thing in vanilla NetHack.
Kinda surprised it isn't. In this commit, if the player or a monster
wields Vorpal Blade, they are protected from being decapitated (any
monster that has AD_HEAD damage type). The feedback is the same as if
they didn't have a head (wildly miss).

Cherry picked from k21971/evilhack@82a54d26b4