Use verbalize for shopkeeper price-check #chatting
The shopkeeper is speaking out loud, so use verbalize for consistency
with other types of speech.

I couldn't figure out a way to wrap the multiline version in quotes in a
way that actually worked and looked good, so I restricted this to the
pline responses.

Cherry picked from nethack/nethack@ee270bfbe0