Feedback from detect unseen and detection
Ported from NetHack 3.7

Secret door detection feedback

When zapping a wand of secret door detection or casting spell of
detect unseem instead of displaying
|You reveal secret doors.
|You reveal a secret corridor.
|You reveal traps.
|You reveal a hidden monster.
show
|You reveal 2 secret doors, a secret corridor, 3 traps, and a
 monster.
as a single message.

Detecting invisible monsters is still a separate message; those

re-mapped as "remembered, unseen monster" but not actually revealed.

nethack/nethack@097d22599

Enhance feedback from detect unseen

I was never too happy with how this was a silent effect that required
you to watch the map to see if anything changed. It might count as
an accessibility issue as well, not sure.

This change adds specific feedback for all the possible things that
might get revealed by detecting unseen. If you reveal a secret door or a
trap, you now get a message indicating that.

One slight behavior change here: if the only thing detected is invisible
monsters, the game previously did not return a result of "detected
a non-zero number of things" to the caller of findit(); now it does.
(This allows the wand to be automatically identified when it prints a
message about detecting invisible monsters.)

nethack/nethack@6645120f2