Add magic platform terrain type
Inspired by copperwater/xnethack@baec1fadd.

"A magic platform is, essentially, a sturdy platform floating in open
air, and would only be expected to appear in conjunction with air.  This
is designed for the new Valkyrie quest, where magic platforms will be
used to span a vast gulf of air."